agendamento e pagamento RabbitMQ

+++ pagamento/src/main/java/com/adjt/pagamento/rest/dto/request/PagamentoRequest.java
@@ -1,3 +1,5 @@
 package com.adjt.pagamento.rest.dto.request;
 
-public record PagamentoRequest(Double valor, String pagamento_id, String cliente_id) {}
+import java.math.BigDecimal;
+
+public record PagamentoRequest(BigDecimal valor, String pagamento_id, String cliente_id) {}
